requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
python-dateutil>=2.8.2

# API clients
requests>=2.31.0
python-dotenv>=1.0.0

# Price data
yfinance>=0.2.36
ccxt>=4.2.0  # Cryptocurrency exchange trading library

# Technical analysis
pandas-ta>=0.3.14b
TA-Lib>=0.4.28  # May need manual installation on some systems

# Sentiment analysis
vaderSentiment>=3.3.2
textblob>=0.17.1
nltk>=3.8.1

# Data visualization
matplotlib>=3.7.0
plotly>=5.18.0
seaborn>=0.13.0

# Configuration
PyYAML>=6.0.1

# Jupyter (optional, for notebooks)
jupyter>=1.0.0
ipykernel>=6.25.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Utilities
tqdm>=4.66.0  # Progress bars
schedule>=1.2.0  # Job scheduling
